move aggregation helper to repo. updated changelog
# Conflicts:
#	CHANGELOG.md
